Check if an egg file named after the first argument exists
Check if an egg file (i.e., .egg file in C5, .setup file in C4) named
after the first argument exists in the egg sources directory fetched
by henrietta-cache.  If it doesn't exist, raise an error.

That's to prevent publishing an egg with a name which is valid for
henrietta-cache but not for chicken-install.

This check was suggested by Sergey Goldgaber.